requirements.txt
geemap
geopandas
earthengine-api
geojson
google-cloud-bigquery
pytest
pytest-mock
PyYAML
db-dtypes
fastapi
uvicorn[standard]
pydantic-settings # Explicitly add for clarity, though often included with pydantic